moved getting orders for dashboard to App component
this way the orders are downloaded only once at the start of the application instead of every time when the user navigates to `/`
