+++ src/components/dashboard.jsx
@@ -3,40 +3,18 @@ import {Row, Col, Alert, Button} from 'react-bootstrap';
 import {Link} from 'react-router';
 import OrderTile from './orderTile';
 import {connect} from 'react-redux';
-import {hydrateOrders} from '../store/ordersActions';
 
 const Dashboard = React.createClass({
     propTypes: {
-        dispatch: React.PropTypes.func.isRequired,
         orders: React.PropTypes.array.isRequired,
         resources: React.PropTypes.object.isRequired
-    },
-
-    componentDidMount () {
-        fetch('/api/orders')
-            .then(response => response.json())
-            .then(orders => {
-                const ordersToday = orders.map(orderWithStringDates => {
-                    return {
-                        ...orderWithStringDates,
-                        deadline: new Date(orderWithStringDates.deadline)
-                    }
-                }).filter(order => {
-                    const today = new Date();
-                    return today.getFullYear() === order.deadline.getFullYear() &&
-                        today.getMonth() === order.deadline.getMonth() &&
-                        today.getDate() === order.deadline.getDate();
-                });
-
-                this.props.dispatch(hydrateOrders(ordersToday));
-            });
     },
 
     _renderOrderTiles () {
         return this.props.orders.map(order => <OrderTile key={order.id} {...order} />);
     },
 
-    _getTileContainerStyles() {
+    _getTileContainerStyles () {
         return {
             marginTop: '1em'
         };
